requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
jinja2>=3.1,<4
pyyaml>=6.0,<7
pydantic>=2.9,<3
pydantic-settings>=2.6,<3
ldap3>=2.9,<3
redis>=5.2,<6
impacket>=0.12,<1
cryptography>=44,<45
python-multipart>=0.0.18,<1
sqlalchemy[asyncio]>=2.0,<3
alembic>=1.14,<2
asyncpg>=0.30,<1
psutil>=6.1,<7
dnspython>=2.7,<3
argon2-cffi>=25,<26
httptools>=0.6,<1
